backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pymongo==4.8.0
boto3==1.35.14
Werkzeug==3.0.3
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
flask-cors==4.0.1
Pillow==10.4.0
tensorflow==2.15.0
numpy==1.26.4
requests==2.32.3
gunicorn==22.0.0
